+++ gcovr/tests/test_gcov_parser.py
@@ -315,6 +315,7 @@ def test_gcov_8(capsys, sourcename):
         filename="tmp.cpp",
         lines=lines,
         exclude_lines_by_pattern=None,
+        exclude_pattern_prefix=None,
         flags=flags,
     )
 
@@ -346,6 +347,7 @@ def test_unknown_tags(caplog, ignore_errors):
             filename="foo.c",
             lines=lines,
             exclude_lines_by_pattern=None,
+            exclude_pattern_prefix=None,
             flags=flags,
         )
 
@@ -384,6 +386,7 @@ def test_pathologic_codeline(caplog):
             filename="foo.c",
             lines=lines,
             exclude_lines_by_pattern=None,
+            exclude_pattern_prefix=None,
             flags=ParserFlags.NONE,
         )
 
@@ -438,6 +441,7 @@ def test_exception_during_coverage_processing(caplog):
                 lines,
                 filename="test.cpp",
                 exclude_lines_by_pattern=None,
+                exclude_pattern_prefix=None,
                 flags=ParserFlags.EXCLUDE_INTERNAL_FUNCTIONS,
             )
 
@@ -488,6 +492,7 @@ def test_trailing_function_tag():
         filename="test.cpp",
         flags=ParserFlags.NONE,
         exclude_lines_by_pattern=None,
+        exclude_pattern_prefix=None,
     )
 
     assert coverage.functions.keys() == {"example"}
@@ -534,6 +539,7 @@ def test_branch_exclusion(flags):
         source.splitlines(),
         filename="example.cpp",
         exclude_lines_by_pattern=None,
+        exclude_pattern_prefix=None,
         flags=flags,
     )
 
@@ -568,6 +574,7 @@ def test_function_exclusion(flags):
         source.splitlines(),
         filename="example.cpp",
         exclude_lines_by_pattern=None,
+        exclude_pattern_prefix=None,
         flags=flags,
     )
 
@@ -600,6 +607,7 @@ def test_noncode_lines():
             flags=flags,
             filename="example.cpp",
             exclude_lines_by_pattern=None,
+            exclude_pattern_prefix=None,
         )
 
         for line_data in coverage.lines.values():